More work on refactoring PlayerTabList
