+++ src/main/java/net/william278/velocitab/vanish/VanishManager.java
@@ -60,7 +60,7 @@ public class VanishManager {
             return;
         }
 
-        plugin.getTabList().vanishPlayer(tabPlayer.get());
+        plugin.getTabList().getVanishTabList().vanishPlayer(tabPlayer.get());
         plugin.getScoreboardManager().ifPresent(scoreboardManager -> scoreboardManager.vanishPlayer(tabPlayer.get()));
     }
 
@@ -71,7 +71,7 @@ public class VanishManager {
             return;
         }
 
-        plugin.getTabList().unVanishPlayer(tabPlayer.get());
+        plugin.getTabList().getVanishTabList().unVanishPlayer(tabPlayer.get());
         plugin.getScoreboardManager().ifPresent(scoreboardManager -> scoreboardManager.unVanishPlayer(tabPlayer.get()));
     }
 }
